+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/update/MediaGroupUpdates/EditMessageMediaGroupUpdate.kt
@@ -2,11 +2,12 @@ package com.github.insanusmokrassar.TelegramBotAPI.types.update.MediaGroupUpdate
 
 import com.github.insanusmokrassar.TelegramBotAPI.types.UpdateIdentifier
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.MediaGroupMessage
+import com.github.insanusmokrassar.TelegramBotAPI.types.update.EditMessageUpdate
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.BaseMessageUpdate
 
 data class EditMessageMediaGroupUpdate(
-    override val origins: List<BaseMessageUpdate>
-) : MediaGroupUpdate {
-    override val updateId: UpdateIdentifier = origins.last().updateId
-    override val data: List<MediaGroupMessage> = origins.mapNotNull { it.data as? MediaGroupMessage }
+    override val origin: EditMessageUpdate
+) : EditMediaGroupUpdate {
+    override val updateId: UpdateIdentifier = origin.updateId
+    override val data: MediaGroupMessage = origin.data as MediaGroupMessage
 }