+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/types/update/MediaGroupUpdates/MediaGroupUpdate.kt
@@ -4,7 +4,14 @@ import com.github.insanusmokrassar.TelegramBotAPI.types.message.abstracts.MediaG
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.BaseMessageUpdate
 import com.github.insanusmokrassar.TelegramBotAPI.types.update.abstracts.Update
 
-interface MediaGroupUpdate: Update {
+interface MediaGroupUpdate : Update
+
+interface SentMediaGroupUpdate: MediaGroupUpdate {
     override val data: List<MediaGroupMessage>
     val origins: List<BaseMessageUpdate>
+}
+
+interface EditMediaGroupUpdate : MediaGroupUpdate {
+    override val data: MediaGroupMessage
+    val origin: BaseMessageUpdate
 }